+++ besu/src/test/java/org/hyperledger/besu/ForkIdsNetworkConfigTest.java
@@ -31,12 +31,11 @@ import org.hyperledger.besu.consensus.merge.TransitionUtils;
 import org.hyperledger.besu.ethereum.chain.Blockchain;
 import org.hyperledger.besu.ethereum.chain.GenesisState;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
-import org.hyperledger.besu.ethereum.core.BlockNumberStreamingProtocolSchedule;
 import org.hyperledger.besu.ethereum.core.MilestoneStreamingProtocolSchedule;
 import org.hyperledger.besu.ethereum.forkid.ForkId;
 import org.hyperledger.besu.ethereum.forkid.ForkIdManager;
+import org.hyperledger.besu.ethereum.mainnet.DefaultProtocolSchedule;
 import org.hyperledger.besu.ethereum.mainnet.MainnetProtocolSchedule;
-import org.hyperledger.besu.ethereum.mainnet.UnifiedProtocolSchedule;
 
 import java.util.Collection;
 import java.util.List;
@@ -210,12 +209,12 @@ public class ForkIdsNetworkConfigTest {
     private static MilestoneStreamingTransitionProtocolSchedule createSchedule(
         final GenesisConfigFile genesisConfigFile) {
       final GenesisConfigOptions configOptions = genesisConfigFile.getConfigOptions();
-      BlockNumberStreamingProtocolSchedule preMergeProtocolSchedule =
-          new BlockNumberStreamingProtocolSchedule(
-              (UnifiedProtocolSchedule) MainnetProtocolSchedule.fromConfig(configOptions));
-      BlockNumberStreamingProtocolSchedule postMergeProtocolSchedule =
-          new BlockNumberStreamingProtocolSchedule(
-              (UnifiedProtocolSchedule) MergeProtocolSchedule.create(configOptions, false));
+      MilestoneStreamingProtocolSchedule preMergeProtocolSchedule =
+          new MilestoneStreamingProtocolSchedule(
+              (DefaultProtocolSchedule) MainnetProtocolSchedule.fromConfig(configOptions));
+      MilestoneStreamingProtocolSchedule postMergeProtocolSchedule =
+          new MilestoneStreamingProtocolSchedule(
+              (DefaultProtocolSchedule) MergeProtocolSchedule.create(configOptions, false));
       final MilestoneStreamingTransitionProtocolSchedule schedule =
           new MilestoneStreamingTransitionProtocolSchedule(
               preMergeProtocolSchedule, postMergeProtocolSchedule);
@@ -228,8 +227,8 @@ public class ForkIdsNetworkConfigTest {
       private final TransitionUtils<MilestoneStreamingProtocolSchedule> transitionUtils;
 
       public MilestoneStreamingTransitionProtocolSchedule(
-          final BlockNumberStreamingProtocolSchedule preMergeProtocolSchedule,
-          final BlockNumberStreamingProtocolSchedule postMergeProtocolSchedule) {
+          final MilestoneStreamingProtocolSchedule preMergeProtocolSchedule,
+          final MilestoneStreamingProtocolSchedule postMergeProtocolSchedule) {
         super(preMergeProtocolSchedule, postMergeProtocolSchedule, PostMergeContext.get());
         transitionUtils =
             new TransitionUtils<>(
