+++ consensus/qbft/src/test/java/org/hyperledger/besu/consensus/qbft/QbftProtocolScheduleTest.java
@@ -33,7 +33,7 @@ import org.hyperledger.besu.cryptoservices.NodeKeyUtils;
 import org.hyperledger.besu.datatypes.Address;
 import org.hyperledger.besu.ethereum.ProtocolContext;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
-import org.hyperledger.besu.ethereum.core.BlockNumberStreamingProtocolSchedule;
+import org.hyperledger.besu.ethereum.core.MilestoneStreamingProtocolSchedule;
 import org.hyperledger.besu.ethereum.core.PrivacyParameters;
 import org.hyperledger.besu.ethereum.core.Util;
 import org.hyperledger.besu.ethereum.mainnet.HeaderValidationMode;
@@ -89,7 +89,7 @@ public class QbftProtocolScheduleTest {
                 new ForkSpec<>(0, qbftConfigOptions),
                 new ForkSpec<>(1, arbitraryTransition),
                 new ForkSpec<>(2, contractTransition)));
-    assertThat(new BlockNumberStreamingProtocolSchedule(schedule).streamMilestoneBlocks().count())
+    assertThat(new MilestoneStreamingProtocolSchedule(schedule).streamMilestoneBlocks().count())
         .isEqualTo(3);
     assertThat(validateHeader(schedule, validators, parentHeader, blockHeader, 0)).isTrue();
     assertThat(validateHeader(schedule, validators, parentHeader, blockHeader, 1)).isTrue();
@@ -118,7 +118,7 @@ public class QbftProtocolScheduleTest {
                 new ForkSpec<>(0, JsonQbftConfigOptions.DEFAULT),
                 new ForkSpec<>(1, arbitraryTransition),
                 new ForkSpec<>(2, JsonQbftConfigOptions.DEFAULT)));
-    assertThat(new BlockNumberStreamingProtocolSchedule(schedule).streamMilestoneBlocks().count())
+    assertThat(new MilestoneStreamingProtocolSchedule(schedule).streamMilestoneBlocks().count())
         .isEqualTo(3);
     assertThat(validateHeader(schedule, validators, parentHeader, blockHeader, 0)).isTrue();
     assertThat(validateHeader(schedule, validators, parentHeader, blockHeader, 1)).isTrue();
