+++ ethereum/blockcreation/src/test/java/org/hyperledger/besu/ethereum/blockcreation/BlockMinerTest.java
@@ -31,11 +31,11 @@ import org.hyperledger.besu.ethereum.core.BlockHeader;
 import org.hyperledger.besu.ethereum.core.BlockHeaderTestFixture;
 import org.hyperledger.besu.ethereum.core.BlockImporter;
 import org.hyperledger.besu.ethereum.mainnet.BlockImportResult;
+import org.hyperledger.besu.ethereum.mainnet.DefaultProtocolSchedule;
 import org.hyperledger.besu.ethereum.mainnet.HeaderValidationMode;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSpec;
 import org.hyperledger.besu.ethereum.mainnet.ScheduledProtocolSpec.BlockNumberProtocolSpec;
-import org.hyperledger.besu.ethereum.mainnet.UnifiedProtocolSchedule;
 import org.hyperledger.besu.util.Subscribers;
 
 import java.math.BigInteger;
@@ -143,8 +143,8 @@ public class BlockMinerTest {
   }
 
   private ProtocolSchedule singleSpecSchedule(final ProtocolSpec protocolSpec) {
-    final UnifiedProtocolSchedule protocolSchedule =
-        new UnifiedProtocolSchedule(Optional.of(BigInteger.valueOf(1234)));
+    final DefaultProtocolSchedule protocolSchedule =
+        new DefaultProtocolSchedule(Optional.of(BigInteger.valueOf(1234)));
     protocolSchedule.putMilestone(BlockNumberProtocolSpec::create, 0, protocolSpec);
     return protocolSchedule;
   }
