+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/mainnet/ProtocolScheduleBuilder.java
@@ -15,8 +15,10 @@
 package org.hyperledger.besu.ethereum.mainnet;
 
 import org.hyperledger.besu.config.GenesisConfigOptions;
+import org.hyperledger.besu.ethereum.chain.BadBlockManager;
 import org.hyperledger.besu.ethereum.core.PrivacyParameters;
 import org.hyperledger.besu.ethereum.mainnet.ScheduledProtocolSpec.BlockNumberProtocolSpec;
+import org.hyperledger.besu.ethereum.privacy.PrivateTransactionValidator;
 import org.hyperledger.besu.evm.internal.EvmConfiguration;
 
 import java.math.BigInteger;
@@ -24,12 +26,24 @@ import java.util.Optional;
 import java.util.OptionalLong;
 import java.util.TreeMap;
 import java.util.function.Function;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public class ProtocolScheduleBuilder extends AbstractProtocolScheduleBuilder {
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+public class ProtocolScheduleBuilder {
+
+  private static final Logger LOG = LoggerFactory.getLogger(ProtocolScheduleBuilder.class);
+  private final GenesisConfigOptions config;
   private final Optional<BigInteger> defaultChainId;
-  private UnifiedProtocolSchedule protocolSchedule;
+  private final ProtocolSpecAdapters protocolSpecAdapters;
+  private final PrivacyParameters privacyParameters;
+  private final boolean isRevertReasonEnabled;
+  private final EvmConfiguration evmConfiguration;
+  private final BadBlockManager badBlockManager = new BadBlockManager();
+
+  private DefaultProtocolSchedule protocolSchedule;
 
   public ProtocolScheduleBuilder(
       final GenesisConfigOptions config,
@@ -69,19 +83,127 @@ public class ProtocolScheduleBuilder extends AbstractProtocolScheduleBuilder {
       final PrivacyParameters privacyParameters,
       final boolean isRevertReasonEnabled,
       final EvmConfiguration evmConfiguration) {
-    super(config, protocolSpecAdapters, privacyParameters, isRevertReasonEnabled, evmConfiguration);
+    this.config = config;
+    this.protocolSpecAdapters = protocolSpecAdapters;
+    this.privacyParameters = privacyParameters;
+    this.isRevertReasonEnabled = isRevertReasonEnabled;
+    this.evmConfiguration = evmConfiguration;
     this.defaultChainId = defaultChainId;
   }
 
   public ProtocolSchedule createProtocolSchedule() {
     final Optional<BigInteger> chainId = config.getChainId().or(() -> defaultChainId);
-    protocolSchedule = new UnifiedProtocolSchedule(chainId);
+    protocolSchedule = new DefaultProtocolSchedule(chainId);
     initSchedule(protocolSchedule, chainId);
     return protocolSchedule;
   }
 
-  @Override
-  protected void validateForkOrdering() {
+  private void initSchedule(
+      final ProtocolSchedule protocolSchedule, final Optional<BigInteger> chainId) {
+
+    final MainnetProtocolSpecFactory specFactory =
+        new MainnetProtocolSpecFactory(
+            chainId,
+            config.getContractSizeLimit(),
+            config.getEvmStackSize(),
+            isRevertReasonEnabled,
+            config.getEcip1017EraRounds(),
+            evmConfiguration);
+
+    validateForkOrdering();
+
+    final TreeMap<Long, BuilderMapEntry> builders = buildMilestoneMap(specFactory);
+
+    // At this stage, all milestones are flagged with correct modifier, but ProtocolSpecs must be
+    // inserted _AT_ the modifier block entry.
+    if (!builders.isEmpty()) {
+      protocolSpecAdapters.stream()
+          .forEach(
+              entry -> {
+                final long modifierBlock = entry.getKey();
+                final BuilderMapEntry parent =
+                    Optional.ofNullable(builders.floorEntry(modifierBlock))
+                        .orElse(builders.firstEntry())
+                        .getValue();
+                builders.put(
+                    modifierBlock,
+                    new BuilderMapEntry(
+                        parent.scheduledSpecFactory,
+                        modifierBlock,
+                        parent.getBuilder(),
+                        entry.getValue()));
+              });
+    }
+
+    // Create the ProtocolSchedule, such that the Dao/fork milestones can be inserted
+    builders
+        .values()
+        .forEach(
+            e ->
+                addProtocolSpec(
+                    protocolSchedule,
+                    e.scheduledSpecFactory,
+                    e.getBlockIdentifier(),
+                    e.getBuilder(),
+                    e.modifier));
+
+    // NOTE: It is assumed that Daofork blocks will not be used for private networks
+    // as too many risks exist around inserting a protocol-spec between daoBlock and daoBlock+10.
+    config
+        .getDaoForkBlock()
+        .ifPresent(
+            daoBlockNumber -> {
+              final BuilderMapEntry previousSpecBuilder =
+                  builders.floorEntry(daoBlockNumber).getValue();
+              final ProtocolSpec originalProtocolSpec =
+                  getProtocolSpec(
+                      protocolSchedule,
+                      previousSpecBuilder.getBuilder(),
+                      previousSpecBuilder.getModifier());
+              addProtocolSpec(
+                  protocolSchedule,
+                  BlockNumberProtocolSpec::create,
+                  daoBlockNumber,
+                  specFactory.daoRecoveryInitDefinition(),
+                  protocolSpecAdapters.getModifierForBlock(daoBlockNumber));
+              addProtocolSpec(
+                  protocolSchedule,
+                  BlockNumberProtocolSpec::create,
+                  daoBlockNumber + 1L,
+                  specFactory.daoRecoveryTransitionDefinition(),
+                  protocolSpecAdapters.getModifierForBlock(daoBlockNumber + 1L));
+              // Return to the previous protocol spec after the dao fork has completed.
+              protocolSchedule.putMilestone(
+                  BlockNumberProtocolSpec::create, daoBlockNumber + 10, originalProtocolSpec);
+            });
+
+    // specs for classic network
+    config
+        .getClassicForkBlock()
+        .ifPresent(
+            classicBlockNumber -> {
+              final BuilderMapEntry previousSpecBuilder =
+                  builders.floorEntry(classicBlockNumber).getValue();
+              final ProtocolSpec originalProtocolSpec =
+                  getProtocolSpec(
+                      protocolSchedule,
+                      previousSpecBuilder.getBuilder(),
+                      previousSpecBuilder.getModifier());
+              addProtocolSpec(
+                  protocolSchedule,
+                  BlockNumberProtocolSpec::create,
+                  classicBlockNumber,
+                  ClassicProtocolSpecs.classicRecoveryInitDefinition(
+                      config.getContractSizeLimit(), config.getEvmStackSize(), evmConfiguration),
+                  Function.identity());
+              protocolSchedule.putMilestone(
+                  BlockNumberProtocolSpec::create, classicBlockNumber + 1, originalProtocolSpec);
+            });
+
+    LOG.info("Protocol schedule created with milestones: {}", protocolSchedule.listMilestones());
+  }
+
+  private void validateForkOrdering() {
     if (config.getDaoForkBlock().isEmpty()) {
       validateClassicForkOrdering();
     } else {
@@ -141,8 +263,31 @@ public class ProtocolScheduleBuilder extends AbstractProtocolScheduleBuilder {
     assert (lastForkBlock >= 0);
   }
 
-  @Override
-  protected Stream<Optional<BuilderMapEntry>> createMilestones(
+  private long validateForkOrder(
+      final String forkName, final OptionalLong thisForkBlock, final long lastForkBlock) {
+    final long referenceForkBlock = thisForkBlock.orElse(lastForkBlock);
+    if (lastForkBlock > referenceForkBlock) {
+      throw new RuntimeException(
+          String.format(
+              "Genesis Config Error: '%s' is scheduled for milestone %d but it must be on or after milestone %d.",
+              forkName, thisForkBlock.getAsLong(), lastForkBlock));
+    }
+    return referenceForkBlock;
+  }
+
+  private TreeMap<Long, BuilderMapEntry> buildMilestoneMap(
+      final MainnetProtocolSpecFactory specFactory) {
+    return createMilestones(specFactory)
+        .flatMap(Optional::stream)
+        .collect(
+            Collectors.toMap(
+                BuilderMapEntry::getBlockIdentifier,
+                b -> b,
+                (existing, replacement) -> replacement,
+                TreeMap::new));
+  }
+
+  private Stream<Optional<BuilderMapEntry>> createMilestones(
       final MainnetProtocolSpecFactory specFactory) {
     return Stream.of(
         blockNumberMilestone(OptionalLong.of(0), specFactory.frontierDefinition()),
@@ -189,60 +334,81 @@ public class ProtocolScheduleBuilder extends AbstractProtocolScheduleBuilder {
         blockNumberMilestone(config.getMystiqueBlockNumber(), specFactory.mystiqueDefinition()));
   }
 
-  @Override
-  protected void postBuildStep(
-      final MainnetProtocolSpecFactory specFactory, final TreeMap<Long, BuilderMapEntry> builders) {
-    // NOTE: It is assumed that Daofork blocks will not be used for private networks
-    // as too many risks exist around inserting a protocol-spec between daoBlock and daoBlock+10.
-    config
-        .getDaoForkBlock()
-        .ifPresent(
-            daoBlockNumber -> {
-              final BuilderMapEntry previousSpecBuilder =
-                  builders.floorEntry(daoBlockNumber).getValue();
-              final ProtocolSpec originalProtocolSpec =
-                  getProtocolSpec(
-                      protocolSchedule,
-                      previousSpecBuilder.getBuilder(),
-                      previousSpecBuilder.getModifier());
-              addProtocolSpec(
-                  protocolSchedule,
-                  BlockNumberProtocolSpec::create,
-                  daoBlockNumber,
-                  specFactory.daoRecoveryInitDefinition(),
-                  protocolSpecAdapters.getModifierForBlock(daoBlockNumber));
-              addProtocolSpec(
-                  protocolSchedule,
-                  BlockNumberProtocolSpec::create,
-                  daoBlockNumber + 1L,
-                  specFactory.daoRecoveryTransitionDefinition(),
-                  protocolSpecAdapters.getModifierForBlock(daoBlockNumber + 1L));
-              // Return to the previous protocol spec after the dao fork has completed.
-              protocolSchedule.putMilestone(
-                  BlockNumberProtocolSpec::create, daoBlockNumber + 10, originalProtocolSpec);
-            });
+  private Optional<BuilderMapEntry> timestampMilestone(
+      final OptionalLong blockIdentifier, final ProtocolSpecBuilder builder) {
+    return createMilestone(
+        blockIdentifier, builder, ScheduledProtocolSpec.TimestampProtocolSpec::create);
+  }
 
-    // specs for classic network
-    config
-        .getClassicForkBlock()
-        .ifPresent(
-            classicBlockNumber -> {
-              final BuilderMapEntry previousSpecBuilder =
-                  builders.floorEntry(classicBlockNumber).getValue();
-              final ProtocolSpec originalProtocolSpec =
-                  getProtocolSpec(
-                      protocolSchedule,
-                      previousSpecBuilder.getBuilder(),
-                      previousSpecBuilder.getModifier());
-              addProtocolSpec(
-                  protocolSchedule,
-                  BlockNumberProtocolSpec::create,
-                  classicBlockNumber,
-                  ClassicProtocolSpecs.classicRecoveryInitDefinition(
-                      config.getContractSizeLimit(), config.getEvmStackSize(), evmConfiguration),
-                  Function.identity());
-              protocolSchedule.putMilestone(
-                  BlockNumberProtocolSpec::create, classicBlockNumber + 1, originalProtocolSpec);
-            });
+  private Optional<BuilderMapEntry> blockNumberMilestone(
+      final OptionalLong blockIdentifier, final ProtocolSpecBuilder builder) {
+    return createMilestone(blockIdentifier, builder, BlockNumberProtocolSpec::create);
+  }
+
+  private Optional<BuilderMapEntry> createMilestone(
+      final OptionalLong blockIdentifier,
+      final ProtocolSpecBuilder builder,
+      final ScheduledSpecFactory factory) {
+    if (blockIdentifier.isEmpty()) {
+      return Optional.empty();
+    }
+    final long blockVal = blockIdentifier.getAsLong();
+    return Optional.of(
+        new BuilderMapEntry(
+            factory, blockVal, builder, protocolSpecAdapters.getModifierForBlock(blockVal)));
+  }
+
+  private ProtocolSpec getProtocolSpec(
+      final ProtocolSchedule protocolSchedule,
+      final ProtocolSpecBuilder definition,
+      final Function<ProtocolSpecBuilder, ProtocolSpecBuilder> modifier) {
+    definition
+        .badBlocksManager(badBlockManager)
+        .privacyParameters(privacyParameters)
+        .privateTransactionValidatorBuilder(
+            () -> new PrivateTransactionValidator(protocolSchedule.getChainId()));
+
+    return modifier.apply(definition).build(protocolSchedule);
+  }
+
+  private void addProtocolSpec(
+      final ProtocolSchedule protocolSchedule,
+      final ScheduledSpecFactory factory,
+      final long blockNumberOrTimestamp,
+      final ProtocolSpecBuilder definition,
+      final Function<ProtocolSpecBuilder, ProtocolSpecBuilder> modifier) {
+
+    protocolSchedule.putMilestone(
+        factory, blockNumberOrTimestamp, getProtocolSpec(protocolSchedule, definition, modifier));
+  }
+
+  private static class BuilderMapEntry {
+    private final ScheduledSpecFactory scheduledSpecFactory;
+    private final long blockIdentifier;
+    private final ProtocolSpecBuilder builder;
+    private final Function<ProtocolSpecBuilder, ProtocolSpecBuilder> modifier;
+
+    public BuilderMapEntry(
+        final ScheduledSpecFactory factory,
+        final long blockIdentifier,
+        final ProtocolSpecBuilder builder,
+        final Function<ProtocolSpecBuilder, ProtocolSpecBuilder> modifier) {
+      this.scheduledSpecFactory = factory;
+      this.blockIdentifier = blockIdentifier;
+      this.builder = builder;
+      this.modifier = modifier;
+    }
+
+    public long getBlockIdentifier() {
+      return blockIdentifier;
+    }
+
+    public ProtocolSpecBuilder getBuilder() {
+      return builder;
+    }
+
+    public Function<ProtocolSpecBuilder, ProtocolSpecBuilder> getModifier() {
+      return modifier;
+    }
   }
 }
